\name{aml}
\docType{data}
\alias{aml}
\alias{leukemia}
\title{Acute Myelogenous Leukemia survival data}
\description{Survival in patients with Acute Myelogenous Leukemia.
  The question at the time was whether the standard course of
chemotherapy should be extended ('maintainance') for additional
cycles.}
\usage{
aml
leukemia
}
\format{
  \tabular{ll}{
    time:\tab survival or censoring time\cr
    status:\tab censoring status\cr
    x: \tab maintenance chemotherapy given? (factor)\cr
  }
}
\note{
  The usage above assumes that the survival package is attached via
  \code{library(survival)} or via another package that depends on
  survival, e.g., mstate.  If not, then either \code{survival::aml}
  can be used or \code{data(cancer, package="survival")}.
}
\source{
  Rupert G. Miller (1997),
  \emph{Survival Analysis}.
  John Wiley & Sons.
  ISBN: 0-471-25218-2.
}
\keyword{datasets}
