% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cindex.R
\name{cindex}
\alias{cindex}
\title{Concordance Index}
\usage{
cindex(pre, object)
}
\arguments{
\item{pre}{A numeric vector of predicted survival time.}

\item{object}{A dataframe or matrix, the first column is observed survival time the second column is survival status.}
}
\value{
The value of C-index.
}
\description{
Compute the C-index in survival analysis.
}
\details{
The most frequently used evaluation metric of survival models is the concordance index (C-index).
It is a measure of rank correlation between predicted risk scores. Only use uncensored observations
}
\examples{
set.seed(1)
# simulate predicted survival time
x1 <- rnorm(100)
# simulate observed survival time
x2 <- rnorm(100)
# simulate observed survival status
x3 <- rep(c(1, 0), 50)
# calculate C-index
cindex(x1, cbind(x2, x3))
}
