% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.learners.R
\name{list.learners}
\alias{list.learners}
\title{List Available Base Learners}
\usage{
list.learners()
}
\value{
The name of all base learners provided in the package
}
\description{
This function lists all base learners provided in the package.
}
\details{
lm1: linear regression

glm1: generalized linear models

glmnet1: Does k-fold cross-validation to chose best alpha and lambda for generalized linear models via penalized maximum likelihood.

glmnet.lasso: LASSO, lambda is chose by k-fold cross-validation for glmnet

glmnet.ridge: Ridge regression, lambda is chose by k-fold cross-validation for glmnet

rpart1: regression tree

lda1: linear discriminant analysis

qda1: quadratic discriminant analysis

KNN1: k-nearest neighbour classification, k is chose by cross-validation

svm1: support vector machine

surv: parametric survival regression

ela1: AFT model with Elasticnet, weights is the weight of survival time

lasso1: AFT model with Lasso, weights is the weight of survival time
}
\examples{
list.learners()
}
