% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTPS.R
\name{MTPS}
\alias{MTPS}
\title{Fit Models using Revised Stacking Algorithm.}
\usage{
MTPS(
  xmat,
  ymat,
  xmat.list,
  family,
  cv = FALSE,
  residual = TRUE,
  nfold = 5,
  method.step1,
  method.step2,
  resid.type = c("deviance", "pearson", "raw"),
  resid.std = FALSE,
  dist1 = NULL,
  weights = NULL,
  weight2 = FALSE
)
}
\arguments{
\item{xmat}{Predictor matrix, each row is an observation vector.}

\item{ymat}{Responses matrix. Quantitative for family = \strong{"gaussian"}. A factor of two levels for family = \strong{"binomial"}. A survival object for family = \strong{"survival"}.}

\item{xmat.list}{The user defines a numeric vector to specify the number of subset columns of the xmat.}

\item{family}{Response type for each response. If all response variable are within the same family it can be "gaussian", "binomial" or "survival", otherwise it is a vector with elements \strong{"gaussian"}, \strong{"binomial"} and \strong{"survival"} to indicate each response family.}

\item{cv}{Logical, indicate if use Cross-Validation Stacking algorithm.}

\item{residual}{Logical, indicate if use Residual Stacking algorithm.}

\item{nfold}{Integer, number of folds for Cross-Validation Stacking algorithm. The default value is 5.}

\item{method.step1}{Base Learners for fitting models in Step 1 of Stacking Algorithm. It can be one base learner function for all outcomes or a list of base learner functions for each outcome. The list of all base learners can be obtained by \code{list.learners()}.}

\item{method.step2}{Base Learners for fitting models in Step 2 of Stacking Algorithm. (see above)}

\item{resid.type}{The residual type for Residual Stacking.}

\item{resid.std}{Logical, whether or not use standardized residual.}

\item{dist1}{Assumed distribution for response variable. If the argument is a

character string, then it is assumed to name an element from

\code{"survreg.distributions"}. These include \strong{"weibull"},

\strong{"exponential"}, \strong{"gaussian"}, \strong{"logistic"},

\strong{"lognormal"} and \strong{"loglogistic"}. Otherwise, it is assumed

to be a user defined list conforming to the format described in

\code{"survreg.distributions"}.}

\item{weights}{Weight for response.}
}
\value{
It returns a MTPS object. It is a list of 4 parameters containing

information about step 1 and step 2 models and the revised stacking algorithm

method.
}
\description{
Fit a model using standard stacking algorithm or revised stacking algorithms to simultaneous predict multiple outcomes.
}
