% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiFit.R
\name{multiFit}
\alias{multiFit}
\title{Fit models on multiple outcomes.}
\usage{
multiFit(
  xmat,
  ymat,
  xmat.list,
  method,
  family,
  dist1 = dist1,
  weights = weights
)
}
\arguments{
\item{xmat}{Matrix of predictors, each row is an observation vector.}

\item{ymat}{Matrix of outcomes.
Quantitative for family = \strong{"gaussian"}.
A factor of two levels for family = \strong{"binomial"}.
A survival object for family = \strong{"survival"}.}

\item{xmat.list}{The user defines a numeric vector to specify the number of subset columns of the xmat.}

\item{method}{Method for fitting models.
It can be one base learner function for all outcomes or a list of base learner functions for each outcome.
The list of all base learners can be obtained by \code{list.learners()}.}

\item{family}{Response type for each response.
If all response variable are within the same family it can be \strong{"gaussian"}, \strong{"binomial"} or \strong{"survival"},
otherwise it is a vector with elements \strong{"gaussian"}, \strong{"binomial"} and \strong{"survival"} to indicate each response family.}

\item{dist1}{Assumed distribution for response variable.
If the argument is a character string, then it is assumed to name an element from survreg.distributions.
These include \strong{"weibull"}, \strong{"exponential"}, \strong{"gaussian"}, \strong{"logistic"},\strong{"lognormal"} and \strong{"loglogistic"}. Otherwise, it is assumed to be a user defined list conforming to the format described in \code{"survreg.distributions"}.}

\item{weights}{Weight for response.}
}
\value{
It returns a multiFit object. It is a list of 6 parameters containing information about the fitted models and fitted values for each outcome.
}
\description{
This function fit individual models to predict each outcome separately.
}
