\name{survival-package}
\alias{survival-package}
\alias{survival}
\docType{package}
\title{
\packageTitle{survival}
}
\description{
\packageDescription{survival}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{survival}
\packageIndices{survival}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{survival}

Maintainer: \packageMaintainer{survival}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
