\name{tt}
\alias{tt}
\title{Mark time tranform terms}
\description{
  This is a dummy function, used to mark terms in the formula
  argument of \code{coxph} for time dependent transformations.
}
\usage{tt(x)}
\arguments{
  \item{x}{a numeric variable}
}
\details{
 This function is not used outside of \code{coxph}, see that help page
 for an example.
}
\value{x, unchanged}
\seealso{\code{\link{coxph}}}
\keyword{ survival }

