\name{ovarian}
\alias{ovarian}
\docType{data}
\title{Ovarian Cancer Survival Data}
\usage{ovarian}
\description{Survival in a randomised trial comparing two treatments for
    ovarian cancer (? from an Eastern Cooperative Oncology Group study)}
\format{
  \tabular{ll}{
    futime:\tab survival or censoring time\cr
    fustat:\tab censoring status\cr
    age: \tab in years\cr
    resid.ds:\tab residual disease present (1=no,2=yes)\cr
    rx:\tab treatment group\cr
    ecog.ps:\tab ECOG performance status (1 is better, see reference)\cr
  }
}
\source{Terry Therneau}
\references{for ECOG performance status: \url{http://ecog.dfci.harvard.edu/~ecogdba/general/perf_stat.html}}
\keyword{datasets}
\keyword{survival}
