\name{rats}
\alias{rats}
\title{Rat data from survival5}
\usage{
data(rats)
}
\format{
\tabular{ll}{
  rat:\tab id\cr
  rx:\tab treatment,(1=drug, 0=control) \cr
  observation:\tab within rat\cr
  start:\tab entry time\cr
  stop:\tab exit time\cr
  status:\tab event status\cr
}
}
\source{Gail, Sautner and Brown, Biometrics 36, 255-66, 1980}

\description{48 rats were injected with a carcinogen, and then
  randomized to either drug or placebo.  The number of tumors ranges
  from 0 to 13; all rats were  censored at 6 months after randomization.
}
\keyword{survival}
\keyword{datasets}
