\name{veteran}
\alias{veteran}
\title{Veterans' Administration Lung Cancer study}
\usage{data(veteran)}
\description{Randomised trial of two treatment regimens for lung
    cancer. This is a standard survival analysis data set}
\format{
    \tabular{ll}{
	trt:\tab 1=standard 2=test\cr
	celltype:\tab 1=squamous,  2=smallcell,  3=adeno,  4=large\cr
	time:\tab survival time\cr
	status:\tab censoring status\cr
	karno:\tab Karnofsky performance score (100=good)\cr
	diagtime:\tab months from diagnosis to randomisation\cr
	age:\tab in years\cr
	prior:\tab prior therapy 0=no, 1=yes\cr
     }
 }
\source{Kalbfleisch and Prentice "The Statistical Analysis of Failure Time Data"}
\keyword{datasets}
